refactor: split output.rs into focused submodules [OPE-483]
Split monolithic output.rs (466 LOC) into four focused modules:
- types.rs: OutputMode enum and CliOutput struct
- table.rs: format_table and push_table_row helpers
- errors.rs: FriendlyError, print_error, print_clap_error
- tests.rs: all 25 unit tests

Public API preserved through mod.rs re-exports.
