+++ crates/opengoose-cli/src/cmd/output/mod.rs
@@ -1,0 +1,10 @@
+mod errors;
+mod table;
+mod types;
+
+#[cfg(test)]
+mod tests;
+
+pub use errors::{print_clap_error, print_error};
+pub use table::format_table;
+pub use types::{CliOutput, OutputMode};
